Player: Store "id" as a size_t rather than string; don't track is_human.
This new "id" is just the index of the object in the "parent" Game's
m_players vector.

Since the game engine should have no concept of how it will receive
input, it should make no difference whether the Player is representing
a human or "CPU" player.  Those sorts of abstraction should be handled
outside the engine itself.  Also for this reason, it makes no sense
for the Player to track some uniquely-identifying name, since all
players are exactly the same to the engine.