Action: Stub out the Action-related structs
These structs will form the building blocks of the interface between
the game engine and the user.  For example, the user will have some
way to query the game engine for a player's "available actions".  The
game engine will then eventually return (possibly over the wire) a
vector of the Action items.

Conversely, when a user wishes for a Player to *take* an action, s/he
will pass an Action struct (possibly over the wire) to the game engine
to actually execute that action, which will mutate a Game object.

Using only enums (backed by ints) and size_t's (also backed by ints)
will hopefully make the communication protocol light.  We shall see
if that actualls ends up being the case :^)