Player: Track current / max number of trade offers per turn
We'll want to be able to enforce a cap on the number of trades that
a player can offer on their turn.  This also means we'll be able to
get away with not *generating every possible trade* when we call into
::get_available_actions() -- we can just limit whether the user is
actually able to offer more trades and leave the particular amounts
up to them.

Maybe we should have a way to prevent duplicate trade offerings? In
that case, we could just change

    size_t m_num_trades_offered_this_turn

into

    std::set<Trade> m_trades_offered_this_turn

and enforce uniqueness that way.