Player: Get rid of Flags calculation
Turns out that we don't actually need to compute all of these things
to do the checks necessary for ::get_available_actions().  In fact,
this method of computation was a holdover from some half-baked idea
I had for CatOnline about *minimizing memory footprint* ... as if
that micro-optimization mattered for JavaScript :^)

It's probably better if we just calculate only the things we need,
as we need them.