Engine: Migrate away from default cases for enum switches
Clang still warns if we return directly from the case statements (since
under the hood it's just desugared to goto's -- I guess clang can't work
with that?), so in some cases we've needed to add assertions.

It's probably cleaner to use the <cassert> lib here instead of <stdexcept>,
since using <stdexcept> gives the impression that we may actually be using
exceptions for control flow.