Result: Add a container for the outcome of Game::execute_action()
The idea is that this struct will hold an error code (or ResultType::Ok)
so that the caller can tell why a given call may not have succeeded. It
probably makes the most sense to just keep these as error codes instead
of std::strings (as originally intended) because std::sprintf() looks
like a nightmare (lmao).

:^)