Game: Add a layer of abstraction under the ::execute() interface
This is really just a partial commit in which both the old abstraction
and the new exist in sequence.  Basically, the Game class shouldn't
be parsing some sort of binary interface when it's trying to execute
actions.  Really, that parsing should happen a layer earlier in the
pipeline when we're first converting incoming data into our own data
structures.  Conceivably, there will be programs that won't need to
rely on explicit binary parsing, since they could link into us and use
the ABI directly.

Anyways, hopefully we can start to clean this stuff up soon :^)