Server: Add a super simple ::write()-based response mechanism
This will just bump the bytes out from the encoded ByteBuffer to the
file descriptor of the socket our request came in on.  We'll need to
be a bit smarter about this eventually (using some sort of Client
abstraction), but this is ok for now.

The good news is that we can now do this:

$ printf '\0\5hello\5world' | nc localhost 6900 | xxd
00000000: 0001 0000 0000 0000 0001 0d25 f38f 322e  ...........%..2.
00000010: dbf8                                     ..