Game: Implement join()ing / leave()ing / start()ing
The basic idea is that once we ::initialize() game, it should be in
a state where different players (identified by their Game::PlayerId)
can join or leave.

Once there are enough players, the game is in a state where it can
be "start"ed, which finishes the initialization by randomizing the
player order and moving the first player to the Vertex::Root.  Once
a game has been started, you can no longer join() or leave() it.