Server: So it turns out that epoll is a bit more complicated ...
It's not as simple as always being able to write directly back to
the socket that we just read from.  Instead, we'll have to push the
responses onto a queue that we then process whenever a client socket
becomes writeable.

In order to actually implement this, we'll need to think through our
Client abstraction a bit.