ServerBase: Improvement epoll()-based socket mgmt
This patch now separates out the reading from the writing, allowing
subclasses to handle those events independently.

The test-server is bit broken at the moment -- it will just try to
write as many bytes as it can to the socket and then loop around and
wait to write more.  A smarter choice would be to have some sort of
basic message queue that it would read from.