mega-commit: Rewrite everything in Rust :^)
This removes all of the C++/Haskell code that I've been fooling around
with, since it feels like writing everything in Rust is going to be a
lot more interesting and productive (than either of those two).

For now, I'm going to keep the old code around, since I'll definitely
need to look back at different bits of the implementation to see how
I've solved problems in the past.

The new code will be broken up into components ("Cargo workspaces"),
with

* `kcore`: the game engine
* `kserver`: a server that manages connections to that engine (TBD)
* `kclient`: one or more clients of `kserver` (TBD)

You can build/test/run the project with

    $ cargo build
    $ cargo test
    $ cargo run